+++ SensorAdventure/src/winlab/sql/SnapShot_SQL.java
@@ -58,7 +58,7 @@ private static class DatabaseHelper extends SQLiteOpenHelper {
 	@Override
 	public void onCreate(SQLiteDatabase db) {
 		for (int i = 0; i < DATABASE_TABLE.length - 1; i++)
-
+		     if ((i<13)&&(Sensors_SQLite_Setting.sensors[i]))
 			switch (i + 1) {
 			case 1:
 			case 2:
@@ -96,10 +96,10 @@ private static class DatabaseHelper extends SQLiteOpenHelper {
 				break;
 
 			}
-		db.execSQL("create table "
+		/*db.execSQL("create table "
 				+ DATABASE_TABLE[DATABASE_TABLE.length - 1] + " ("
 				+ KEY_ROWID + " integer primary key autoincrement, "
-				+ KEY_SAMPLE + " blob not null);");
+				+ KEY_SAMPLE + " blob not null);");*/
 	}
 
 	@Override
@@ -125,6 +125,7 @@ public void close() {
 
 public void deleteTable() {
 	for (int i = 0; i < DATABASE_TABLE.length; i++)
+	     if ((i<13)&&(Sensors_SQLite_Setting.sensors[i]))
 		db.delete(DATABASE_TABLE[i], null, null);
 }
